test: split daemon channel surface roots by format
