+++ crates/daemon/tests/integration/mod.rs
@@ -5,7 +5,8 @@ use super::*;
 mod acp;
 mod architecture;
 mod ask_cli;
-mod channel_surfaces;
+mod channel_surfaces_json;
+mod channel_surfaces_text;
 mod chat_cli;
 mod cli_tests;
 mod doctor_feishu;
